+++ utils/utils.h
@@ -3,7 +3,7 @@
 
 # include "./../push_swap.h"
 
-int		ft_atoi(const char *str);
+long int	ft_atoi(const char *str);
 void	ft_free(void *ptr);
 
 // Listas
@@ -12,7 +12,7 @@ void	ft_lstadd_front(t_stack **lst, t_stack *new);
 void	ft_lstclear(t_stack **lst);
 void	ft_lstdelone(t_stack *lst);
 t_stack	*ft_lstlast(t_stack *lst);
-t_stack	*ft_lstnew(int number);
+t_stack	*ft_lstnew(long int number);
 int		ft_lstsize(t_stack *lst);
 
 /**
